added new picteres and refactored sagas dish component(added toastr function)

+++ src/sagas/dishes.js
@@ -1,6 +1,8 @@
 import 'babel-polyfill';
 import { takeLatest } from 'redux-saga/effects';
 import { put, call, select } from 'redux-saga/effects';
+import { push } from 'react-router-redux';
+import {actions as toastrActions} from 'react-redux-toastr';
 import { GET_DISHES, DELETE_DISH, POST_DISH } from '../constants/dishes';
 import {
 	getDishesSuccess,
@@ -10,6 +12,8 @@ import {
 	postDishSuccess,
 	postDishFailure,
 } from '../actions/dishes';
+
+import { logoutUser } from '../actions/auth';
 
 const selectedDish = (state) => {
     return state.getIn(['dishes', 'list']).toJS();
@@ -29,24 +33,36 @@ const fetchDishes = () => {
 };
 
 const deleteServerDish = (id) => {
-    return fetch(`http://localhost:8080/dishes/${id}`, {
-        headers: new Headers({
-            'Content-Type': 'application/json',
-        }),
-        method: 'DELETE',
-    })
-        .then(response => response.json());
+	return fetch(`http://localhost:8080/dishes/${id}`, {
+		headers: new Headers({
+			'Content-Type': 'application/json',
+			'x-access-token': localStorage.getItem('token')
+		}),
+		method: 'DELETE',
+	})
+		.then(response => {
+			if (response.status === 200) {
+				return response.json();
+			}
+			throw response.json();
+		});
 };
 
 const postServerDish = (dish) => {
 	return fetch('http://localhost:8080/dishes', {
 		headers: new Headers({
-			'Content-Type': 'application/json'
+			'Content-Type': 'application/json',
+			'x-access-token': localStorage.getItem('token')
 		}),
 		method: 'POST',
 		body: JSON.stringify(dish)
 	})
-		.then(res => res.json());
+		.then(res => {
+			if (res.status === 200) {
+				return res.json();
+			}
+			throw res;
+		});
 };
 
 function* getDishes () {
@@ -59,14 +75,31 @@ function* getDishes () {
 }
 
 function* deleteDish (action) {
-    const { id } = action;
-    const dishes = yield select(selectedDish);
-    try {
-        yield call(deleteServerDish, id);
-        yield put(deleteDishSuccess(dishes.filter(dish => dish._id !== id)));
-    } catch (err) {
-        yield put(deleteDishFailure());
-    }
+	const {id} = action;
+	const dishes = yield select(selectedDish);
+	try {
+		const result = yield call(deleteServerDish, id);
+		yield put(toastrActions.add({
+			type: 'success',
+			title: result.message
+		}));
+		yield put(deleteDishSuccess(dishes.filter(dish => dish._id !== id)));
+	} catch (err) {
+		let message = '';
+		if (err.status === 403) {
+			yield put(logoutUser());
+			message = 'Invalid token. You are being logged off';
+		} else {
+			yield put(deleteDishFailure());
+			message = 'Sorry, an error occured!';
+		}
+		localStorage.removeItem('token');
+		yield put(toastrActions.add({
+			type: 'error',
+			title: 'Poket Cookbook',
+			message: message
+		}));
+	}
 }
 
 const getDishForm = (state) => {
@@ -76,13 +109,31 @@ const getDishForm = (state) => {
 function* postDish () {
 	const srcImage = yield select(selectedPicture);
 	const dish = yield select(getDishForm);
-
-	const newDish = Object.assign({}, { srcImage }, dish.values);
+	const newDish = Object.assign({}, {srcImage}, dish.values);
 	try {
-		yield call(postServerDish, newDish);
+		const result = yield call(postServerDish, newDish);
+		yield put(toastrActions.add({
+			type: 'success',
+			title: 'Poket Cookbook',
+			message: result.message
+		}));
 		yield put(postDishSuccess());
+		yield put(push('/admin/dishes'))
 	} catch (err) {
-		yield put(postDishFailure());
+		let message = '';
+		if (err.status === 403) {
+			yield put(logoutUser());
+			message = 'Invalid token. You are being logged off';
+		} else {
+			yield put(postDishFailure());
+			message = 'Sorry, an error occured!';
+		}
+		localStorage.removeItem('token');
+		yield put(toastrActions.add({
+			type: 'error',
+			title: 'Poket Cookbook',
+			message: message
+		}));
 	}
 }
